Added additional trader management logic. Added more ways to stack historical data in . Implemented the .

+++ src/Limitless/Limitless/IMarketBehavior.cs
@@ -4,8 +4,8 @@ namespace Limitless
 {
     public interface IMarketBehavior
     {
-        Task<IOrder?> Buy(string symbol, int quantity, DateTime timestamp, decimal estimatedPrice);
-        Task<IOrder?> Sell(string symbol, int quantity, DateTime timestamp, decimal estimatedPrice);
+        Task<IOrder?> Buy(string symbol, decimal quantity, DateTime timestamp, decimal estimatedPrice);
+        Task<IOrder?> Sell(string symbol, decimal quantity, DateTime timestamp, decimal estimatedPrice);
         Task<IOrder?> GetOrder(IOrder order);
     }
 }
